requirements.txt
deepgram-sdk
langchain
langchain-core
langchain-community
langchain-groq
pypdf
